refactor(qmclient): 整理组件目录模块化
QmClient 组件目录已从平铺结构拆出 voice、translate、hud_notifications 和 monitoring 分组，降低根目录噪声并保持对外组件入口稳定。

## refactor

- 移动 voice、translate、hud_notifications、monitoring 相关源码到子目录

- 同步 CMake 源文件列表、gameclient.h 和测试 include 路径

- 保留 perf_logging.h 在 qmclient 根目录作为跨组件共享工具

## docs

- 增加 components-modularization 探索记录，说明迁移边界和后续不混入菜单拆分

## test

- qmclient_scripts/cmake-windows.cmd --build cmake-build-release --target run_cxx_tests -j 14

- python qmclient_scripts/gate/check_gate.py --mode quick --report-json-path tmp/check-gate-precommit-report.json
