+++ src/game/client/components/qmclient/hud_notifications/hud_notification_catalog.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATION_CATALOG_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATION_CATALOG_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATIONS_HUD_NOTIFICATION_CATALOG_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATIONS_HUD_NOTIFICATION_CATALOG_H
 
 #include "hud_notification_rules.h"
