+++ src/game/client/components/qmclient/hud_notifications/hud_notification_rules.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATION_RULES_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATION_RULES_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATIONS_HUD_NOTIFICATION_RULES_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATIONS_HUD_NOTIFICATION_RULES_H
 
 #include <base/system.h>
