+++ src/game/client/components/qmclient/hud_notifications/hud_notification_static_alias_rules.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATION_STATIC_ALIAS_RULES_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATION_STATIC_ALIAS_RULES_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATIONS_HUD_NOTIFICATION_STATIC_ALIAS_RULES_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_HUD_NOTIFICATIONS_HUD_NOTIFICATION_STATIC_ALIAS_RULES_H
 
 #define QM_HUD_NOTIFICATION_STATIC_ALIAS_RULES(X) \
 	X("你现在会收到私聊消息", WhispersOn) \
