+++ src/game/client/components/qmclient/hud_notifications/hud_notifications.cpp
@@ -1,4 +1,5 @@
 #include "hud_notifications.h"
+
 #include <base/color.h>
 #include <base/system.h>
