+++ src/game/client/components/qmclient/translate/translate.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_H
 
 #include "translate_parse.h"
