+++ src/game/client/components/qmclient/translate/translate_parse.cpp
@@ -1,6 +1,7 @@
 #include "translate_parse.h"
 
 #include <base/system.h>
+
 #include <engine/shared/json.h>
 
 bool ParseLlmResponseJson(const json_value *pObj, SLlmParseResult &Out)
