+++ src/game/client/components/qmclient/translate/translate_parse.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_PARSE_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_PARSE_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_PARSE_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_PARSE_H
 
 // LLM 响应解析结果
 struct SLlmParseResult
@@ -17,4 +17,4 @@ struct _json_value;
 // 返回: true 表示解析成功，false 表示解析失败
 bool ParseLlmResponseJson(const struct _json_value *pObj, SLlmParseResult &Out);
 
-#endif // GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_PARSE_H
+#endif // GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_PARSE_H
