+++ src/game/client/components/qmclient/translate/translate_ui_settings.h
@@ -1,0 +1,12 @@
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_UI_SETTINGS_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_TRANSLATE_TRANSLATE_UI_SETTINGS_H
+
+class CMenus;
+class CUIRect;
+
+namespace NTranslateUiSettings
+{
+	void RenderTranslateUiModule(CMenus *pMenus, CUIRect &CardContent, float LineHeight, float BodySize, float LineSpacing);
+}
+
+#endif
