+++ src/game/client/components/qmclient/voice/voice_capture_pipeline.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_CAPTURE_PIPELINE_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_CAPTURE_PIPELINE_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_CAPTURE_PIPELINE_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_CAPTURE_PIPELINE_H
 
 #include "voice_core.h"
