+++ src/game/client/components/qmclient/voice/voice_component.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_COMPONENT_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_COMPONENT_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_COMPONENT_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_COMPONENT_H
 
 #include "voice_core.h"
 
@@ -33,4 +33,4 @@ private:
 	CVoiceOverlayState m_OverlayState;
 };
 
-#endif // GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_COMPONENT_H
+#endif // GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_COMPONENT_H
