+++ src/game/client/components/qmclient/voice/voice_core.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_CORE_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_CORE_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_CORE_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_CORE_H
 
 #include "voice_utils.h"
 
@@ -339,4 +339,4 @@ public:
 	bool IsOutputUnavailable() const { return m_OutputUnavailable.load(); }
 };
 
-#endif // GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_CORE_H
+#endif // GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_CORE_H
