+++ src/game/client/components/qmclient/voice/voice_utils.h
@@ -1,5 +1,5 @@
-#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_UTILS_H
-#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_UTILS_H
+#ifndef GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_UTILS_H
+#define GAME_CLIENT_COMPONENTS_QMCLIENT_VOICE_VOICE_UTILS_H
 
 #include <base/types.h>
 #include <base/vmath.h>
