+++ src/game/client/gameclient.h
@@ -72,14 +72,14 @@
 #include "components/qmclient/axiom_auto_login.h"
 #include "components/qmclient/collision_hitbox.h"
 #include "components/qmclient/data_version.h"
-#include "components/qmclient/hud_notifications.h"
+#include "components/qmclient/hud_notifications/hud_notifications.h"
 #include "components/qmclient/input_overlay.h"
 #include "components/qmclient/lyrics_component.h"
-#include "components/qmclient/monitoring.h"
+#include "components/qmclient/monitoring/monitoring.h"
 #include "components/qmclient/qmclient.h"
 #include "components/qmclient/scripting.h"
-#include "components/qmclient/translate.h"
-#include "components/qmclient/voice_component.h"
+#include "components/qmclient/translate/translate.h"
+#include "components/qmclient/voice/voice_component.h"
 #include "components/qmclient/weapon_trajectory.h"
 #include "components/race_demo.h"
 #include "components/scoreboard.h"
